+++ spiNNakerGraphFrontEnd.cfg
@@ -93,7 +93,6 @@ reset_machine_on_startup = False
 # BasicRoutingInfoAllocator, BasicDijkstraRouting,
 # MallocBasedRoutingInfoAllocator, GraphEdgeFilter, EdgeToNKeysMapper
 algorithms = BasicDijkstraRouting,RadialPlacer,BasicTagAllocator,MallocBasedRoutingInfoAllocator
-interface_algorithms = FrontEndCommomLoadExecutableImages,FrontEndCommonRoutingTableLoader,FrontEndCommonTagsLoader,FrontEndCommonBufferManagerCreater
 # format is <path1>,<path2>
 extra_xmls_paths = None
 partitionerAlgorithm = PartitionAndPlacePartitioner
